My 3rd rewrite. Basic function doing the "Proper" schema.
