+++ init.lua
@@ -1,6 +1,4 @@
-require "user.options"
-require "user.keymaps"
-require "user.lazy"
-require "user.autocommands"
-require "user.lsp"
-vim.cmd("colorscheme tokyonight-night")
+require "options"
+require "keymaps"
+require "Lazy"
+require "autocommands"
