+++ lua/lsp/mason.lua
@@ -1,0 +1,6 @@
+local lspconfig_status_ok, lspconfig = pcall(require, "lspconfig")
+if not lspconfig_status_ok then
+  return
+end
+
+local opts = {}
