+++ lua/user/autopairs.lua
@@ -1,10 +1,20 @@
--- Setup nvim-cmp.
-local status_ok, npairs = pcall(require, "nvim-autopairs")
-if not status_ok then
-  return
-end
+local M = {
+  "windwp/nvim-autopairs",
+  commit = "0e065d423f9cf649e1d92443c939a4b5073b6768",
+  event = "InsertEnter",
+  dependencies = {
+    {
+      "hrsh7th/nvim-cmp",
+      commit = "cfafe0a1ca8933f7b7968a287d39904156f2c57d",
+      event = {
+        "InsertEnter",
+        "CmdlineEnter",
+      },
+    },
+  },
+}
 
-npairs.setup {
+M.opts = {
   check_ts = true, -- treesitter integration
   disable_filetype = { "TelescopePrompt" },
   ts_config = {
@@ -12,7 +22,6 @@ npairs.setup {
     javascript = { "string", "template_string" },
     java = false,
   },
-
   fast_wrap = {
     map = "<M-e>",
     chars = { "{", "[", "(", '"', "'" },
@@ -26,9 +35,11 @@ npairs.setup {
   },
 }
 
-local cmp_autopairs = require "nvim-autopairs.completion.cmp"
-local cmp_status_ok, cmp = pcall(require, "cmp")
-if not cmp_status_ok then
-  return
+function M.config()
+  local cmp_autopairs = require "nvim-autopairs.completion.cmp"
+  local cmp = require "cmp"
+
+  cmp.event:on("confirm_done", cmp_autopairs.on_confirm_done {})
 end
-cmp.event:on("confirm_done", cmp_autopairs.on_confirm_done {})
+
+return M
