+++ lua/user/gitsigns.lua
@@ -1,9 +1,10 @@
-local status_ok, gitsigns = pcall(require, "gitsigns")
-if not status_ok then
-  return
-end
+local M = {
+  "lewis6991/gitsigns.nvim",
+  commit = "ec4742a7eebf68bec663041d359b95637242b5c3",
+  event = "BufReadPre",
+}
 
-gitsigns.setup {
+M.opts = {
   signs = {
     add = { hl = "GitSignsAdd", text = "▎", numhl = "GitSignsAddNr", linehl = "GitSignsAddLn" },
     change = { hl = "GitSignsChange", text = "▎", numhl = "GitSignsChangeNr", linehl = "GitSignsChangeLn" },
@@ -34,3 +35,5 @@ gitsigns.setup {
     col = 1,
   },
 }
+
+return M
