+++ lua/user/indentline.lua
@@ -1,9 +1,10 @@
-local status_ok, indent_blankline = pcall(require, "indent_blankline")
-if not status_ok then
-  return
-end
+local M = {
+  "lukas-reineke/indent-blankline.nvim",
+  commit = "8299fe7703dfff4b1752aeed271c3b95281a952d",
+  event = "BufReadPre",
+}
 
-indent_blankline.setup {
+M.opts = {
   char = "▏",
   show_trailing_blankline_indent = false,
   show_first_indent_level = true,
@@ -16,3 +17,5 @@ indent_blankline.setup {
     "NvimTree",
   },
 }
+
+return M
