--- lua/user/lsp/init.lua
@@ -1,1 +1,0 @@
-require("user.lsp.handlers")
