+++ lua/user/null-ls.lua
@@ -1,0 +1,37 @@
+local M = {
+  "jose-elias-alvarez/null-ls.nvim",
+  event = "BufReadPre",
+  commit = "60b4a7167c79c7d04d1ff48b55f2235bf58158a7",
+  dependencies = {
+    {
+      "nvim-lua/plenary.nvim",
+      commit = "9a0d3bf7b832818c042aaf30f692b081ddd58bd9",
+      lazy = true,
+    },
+  },
+}
+
+function M.config()
+  local null_ls = require "null-ls"
+  -- https://github.com/jose-elias-alvarez/null-ls.nvim/tree/main/lua/null-ls/builtins/formatting
+  local formatting = null_ls.builtins.formatting
+  -- https://github.com/jose-elias-alvarez/null-ls.nvim/tree/main/lua/null-ls/builtins/diagnostics
+  local diagnostics = null_ls.builtins.diagnostics
+
+  -- https://github.com/prettier-solidity/prettier-plugin-solidity
+  null_ls.setup {
+    debug = false,
+    sources = {
+      formatting.prettier.with {
+        extra_filetypes = { "toml" },
+        extra_args = { "--no-semi", "--single-quote", "--jsx-single-quote" },
+      },
+      formatting.black.with { extra_args = { "--fast" } },
+      formatting.stylua,
+      formatting.google_java_format,
+      diagnostics.flake8,
+    },
+  }
+end
+
+return M
