+++ lua/user/nvim-tree.lua
@@ -1,66 +1,64 @@
-local status_ok, nvim_tree = pcall(require, "nvim-tree")
-if not status_ok then
-  return
-end
-
-local config_status_ok, nvim_tree_config = pcall(require, "nvim-tree.config")
-if not config_status_ok then
-  return
-end
-
-local tree_cb = nvim_tree_config.nvim_tree_callback
-
-nvim_tree.setup {
-  update_focused_file = {
-    enable = true,
-    update_cwd = true,
-  },
-  renderer = {
-    icons = {
-      glyphs = {
-        default = "",
-        symlink = "",
-        folder = {
-          arrow_open = "",
-          arrow_closed = "",
-          default = "",
-          open = "",
-          empty = "",
-          empty_open = "",
-          symlink = "",
-          symlink_open = "",
-        },
-        git = {
-          unstaged = "",
-          staged = "S",
-          unmerged = "",
-          renamed = "➜",
-          untracked = "U",
-          deleted = "",
-          ignored = "◌",
-        },
-      },
-    },
-  },
-  diagnostics = {
-    enable = true,
-    show_on_dirs = true,
-    icons = {
-      hint = "",
-      info = "",
-      warning = "",
-      error = "",
-    },
-  },
-  view = {
-    width = 30,
-    side = "left",
-    mappings = {
-      list = {
-        { key = { "l", "<CR>", "o" }, cb = tree_cb "edit" },
-        { key = "h", cb = tree_cb "close_node" },
-        { key = "v", cb = tree_cb "vsplit" },
-      },
-    },
-  },
+local M = {
+  "kyazdani42/nvim-tree.lua",
+  commit = "59e65d88db177ad1e6a8cffaafd4738420ad20b6",
 }
+
+function M.config()
+  local tree_cb = require("nvim-tree").config.nvim_tree_callback
+  require("nvim-tree").setup {
+    update_focused_file = {
+      enable = true,
+      update_cwd = true,
+    },
+    renderer = {
+      icons = {
+        glyphs = {
+          default = "",
+          symlink = "",
+          folder = {
+            arrow_open = "",
+            arrow_closed = "",
+            default = "",
+            open = "",
+            empty = "",
+            empty_open = "",
+            symlink = "",
+            symlink_open = "",
+          },
+          git = {
+            unstaged = "",
+            staged = "S",
+            unmerged = "",
+            renamed = "➜",
+            untracked = "U",
+            deleted = "",
+            ignored = "◌",
+          },
+        },
+      },
+    },
+    diagnostics = {
+      enable = true,
+      show_on_dirs = true,
+      icons = {
+        hint = "",
+        info = "",
+        warning = "",
+        error = "",
+      },
+    },
+    view = {
+      width = 30,
+      side = "left",
+      mappings = {
+        list = {
+          { key = { "l", "<CR>", "o" }, cb = tree_cb "edit" },
+          { key = "h", cb = tree_cb "close_node" },
+          { key = "v", cb = tree_cb "vsplit" },
+        },
+      },
+    },
+  }
+end
+
+return M
