+++ lua/user/telescope.lua
@@ -1,18 +1,24 @@
-local status_ok, telescope = pcall(require, "telescope")
-if not status_ok then
-  return
-end
+local M = {
+  "nvim-telescope/telescope.nvim",
+  commit = "203bf5609137600d73e8ed82703d6b0e320a5f36",
+  event = "Bufenter",
+  cmd = { "Telescope" },
+  dependencies = {
+    {
+      "ahmedkhalf/project.nvim",
+      commit = "685bc8e3890d2feb07ccf919522c97f7d33b94e4",
+    },
+  },
+}
 
 local actions = require "telescope.actions"
 
-telescope.setup {
+M.opts = {
   defaults = {
-
     prompt_prefix = " ",
     selection_caret = " ",
     path_display = { "smart" },
     file_ignore_patterns = { ".git/", "node_modules" },
-
     mappings = {
       i = {
         ["<Down>"] = actions.cycle_history_next,
@@ -23,3 +29,5 @@ telescope.setup {
     },
   },
 }
+
+return M
