+++ lua/utils/init.lua
@@ -1,0 +1,14 @@
+local M = {}
+
+M.servers = {
+  "lua_ls",
+  "cssls",
+  "html",
+  "tsserver",
+  "pyright",
+  "bashls",
+  "jsonls",
+  "yamlls",
+}
+
+return M
